+++ StrinovaNodeSwitcher.sln
@@ -9,8 +9,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Sources", "Sources", "{5482
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{EC308FD5-8569-430F-AFE7-9E0B846822EB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestCore", "TestCore\TestCore.csproj", "{F52A6097-FF0A-4AA5-8A3A-D44F33663299}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WpfApp", "WpfApp\WpfApp.csproj", "{6E15C4CF-AA74-4DFF-B5C6-ABCC4C0464AA}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoreTests", "CoreTests\CoreTests.csproj", "{725E688F-B0F3-4A55-80D9-6DFDAC5E536E}"
@@ -25,10 +23,6 @@ Global
 		{D0133C68-6938-45E4-9B43-FB91474176AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D0133C68-6938-45E4-9B43-FB91474176AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D0133C68-6938-45E4-9B43-FB91474176AE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F52A6097-FF0A-4AA5-8A3A-D44F33663299}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F52A6097-FF0A-4AA5-8A3A-D44F33663299}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F52A6097-FF0A-4AA5-8A3A-D44F33663299}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F52A6097-FF0A-4AA5-8A3A-D44F33663299}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6E15C4CF-AA74-4DFF-B5C6-ABCC4C0464AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6E15C4CF-AA74-4DFF-B5C6-ABCC4C0464AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6E15C4CF-AA74-4DFF-B5C6-ABCC4C0464AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -43,8 +37,8 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{D0133C68-6938-45E4-9B43-FB91474176AE} = {5482C17D-C8B2-4DE3-A6CD-27BA81BE83B6}
-		{F52A6097-FF0A-4AA5-8A3A-D44F33663299} = {EC308FD5-8569-430F-AFE7-9E0B846822EB}
 		{6E15C4CF-AA74-4DFF-B5C6-ABCC4C0464AA} = {5482C17D-C8B2-4DE3-A6CD-27BA81BE83B6}
+		{725E688F-B0F3-4A55-80D9-6DFDAC5E536E} = {EC308FD5-8569-430F-AFE7-9E0B846822EB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2DAA2811-E5E1-44D1-B3B8-9599D71212A1}
